Refactor: reorganize comparers files
