--- src/PrimeFuncPack.Core/PrimeFuncPack.Core.Box/Internal/BoxEqualityComparerDefault.cs
@@ -1,9 +1,0 @@
-﻿#nullable enable
-
-namespace PrimeFuncPack
-{
-    internal static class BoxEqualityComparerDefault<T>
-    {
-        public static readonly BoxEqualityComparer<T> Value = new BoxEqualityComparer<T>();
-    }
-}
