--- src/PrimeFuncPack.Core/PrimeFuncPack.Core.Box/Internal/BoxSamenessComparerDefault.cs
@@ -1,9 +1,0 @@
-﻿#nullable enable
-
-namespace PrimeFuncPack
-{
-    internal static class BoxSamenessComparerDefault<T>
-    {
-        public static readonly BoxSamenessComparer<T> Value = new BoxSamenessComparer<T>();
-    }
-}
