--- src/PrimeFuncPack.Core/PrimeFuncPack.Core.Optional/Internal/OptionalEqualityComparerDefault.cs
@@ -1,9 +1,0 @@
-﻿#nullable enable
-
-namespace PrimeFuncPack
-{
-    internal static class OptionalEqualityComparerDefault<T>
-    {
-        public static readonly OptionalEqualityComparer<T> Value = new OptionalEqualityComparer<T>();
-    }
-}
