--- src/PrimeFuncPack.Core/PrimeFuncPack.Core.Optional/Internal/OptionalSamenessComparerDefault.cs
@@ -1,9 +1,0 @@
-﻿#nullable enable
-
-namespace PrimeFuncPack
-{
-    internal static class OptionalSamenessComparerDefault<T>
-    {
-        public static readonly OptionalSamenessComparer<T> Value = new OptionalSamenessComparer<T>();
-    }
-}
